requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0

# Web framework
flask>=2.2.0
flask-cors>=3.0.10
flask-limiter>=3.5.0

# Authentication & Security
PyJWT>=2.6.0
bcrypt>=4.0.0

# Dashboard dependencies
Jinja2>=3.1.2

# Data handling
pandas>=1.5.0

# Security
cryptography>=38.0.0

# Logging and monitoring
python-dotenv>=0.20.0

# Development
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.10.0
pylint>=2.15.0
